fix initial "x-show" from LiveComponent
Adding an "id" attr on the element with "x-data" seems to resolve the issue.
Still unclear why the issue persists when an LC auto-generated "id"
is present on the mentioned HTML element.
Moreover, adding a similarly formatted "id", such as "phx-F0_5AaO4bEUTIwXl-1-0", also fixes the issue—but an auto-generated one does _not_.